requirements.txt
Django==1.3.1
PIL==1.1.7
South==0.7.3
boto==2.1.1
django-extensions==0.7.1
django-imagekit==1.1
django-oembed==0.1.3
django-storages==1.1.4
django-taggit==0.9.3
docutils==0.8.1
wsgiref==0.1.2
